Crate some_macros

This crate provides some macros

The use-std feature enables the following macros:

vec_deque

hash_map

hash_set

dbgr

Macros

alt
A macro that returns the first set of token trees it gets

count_args
Counts how many arguments it receives.

dbgr
A macro to print pretty looking and informative debug logs.

debug_lvl
Returns a boolean that tells you whether a specific environment variable had a certain value at compiletime

hash_map
Constructs a HashMap, calling into on the keys and values

hash_set
Constructs a HashSet, calling into on the values

vec_deque
Constructs a VecDeque, calling into on the keys and values