Macro some_macros::dbgr

macro_rules! dbgr {
    () => { ... };
    ($l:literal) => { ... };
    (# $l:literal) => { ... };
    (: $l:literal) => { ... };
    ($($e:expr),+) => { ... };
    (#; $($e:expr),+) => { ... };
}

A macro to print pretty looking and informative debug logs.

Difference from dbg!

There are three differences:

dbgr! takes its arguments by ref (like println!) so it can be called at any point without worrying about ownership

dbgr! has more functionality

Motivation

The intended use case for dbgr! is a different one than that of dbg!.
Where dbg! aims to let you put it in the middle of function calls (like f(dbg!(a), b, c)), dbgr! aims to be more like println!.

Syntax

dbgr! takes zero or more expressions.

dbgr!();     // prints "[<file>] reached <line>:<column>."
dbgr!(a);    // prints "[<file>@<line>] a = <a, debug printed>"
dbgr!(a, b); // prints "[<file>@<line>] a = <a, debug printed>, b = <b, debug printed>"
dbgr!(#; a); // prints "[<file>@<line>] a = <a, debug pretty printed>"
dbgr!(#; a, b); // prints "[<file>@<line>] a = <a, debug pretty printed>, b = <b, debug pretty printed>"
dbgr!(<any literal>); // prints "[<file>@<line>] <the literal, debug printed>" (only works with a single literal)
dbgr!(# <any literal>); // prints "[<file>@<line>] <the literal, debug pretty printed>" (only works with a single literal)
dbgr!(: <any literal>); // prints "[<file>@<line>] <the literal, display printed>" (only works with a single literal)