Macro some_macros::debug_lvl

macro_rules! debug_lvl {
    (@tok >= $a:expr, $b:expr) => { ... };
    (@tok <= $a:expr, $b:expr) => { ... };
    (@tok == $a:expr, $b:expr) => { ... };
    (@tok != $a:expr, $b:expr) => { ... };
    (@tok > $a:expr, $b:expr) => { ... };
    (@tok < $a:expr, $b:expr) => { ... };
    (@tok $t:tt $a:expr, $b:expr) => { ... };
    ($debug_env:literal $t:tt $min:literal) => { ... };
}

Returns a boolean that tells you whether a specific environment variable had a certain value at compiletime

Motivation

This is useful for when you want to easily switch debug code on and off with wnvironment variables. Instead of writing the following code snippet every time

match option_env!(LIT) {
    Some(s) => {
        s.parse::<i64>().map_or(false, |lvl| lvl >= MIN)
    }
    None => false,
}

you can just write debug_lvl!(LIT >= MIN).

This works with all comparison operators (==, !=, >, <, >=, <=)